Generating 2d object set GUI.
Not tested

+++ BinPackingProblem/Logic.ObjectGenerator/ObjectSetProperties.cs
@@ -11,7 +11,7 @@ namespace Logic.ObjectGenerator
     /// </summary>
     public class ObjectSetProperties2D
     {
-        public decimal ObjectAmount { get; set; }
+        public int ObjectAmount { get; set; }
         public decimal AverageObjectWidthHeightRatio { get; set; }
         public decimal ObjectWidthHeightRatioStandardDeviation { get; set; }
         public decimal MinObjectWidthHeightRatio { get; set; }
@@ -27,7 +27,7 @@ namespace Logic.ObjectGenerator
     /// </summary>
     public class ObjectSetProperties3D
     {
-        public decimal ObjectAmount { get; set; }
+        public int ObjectAmount { get; set; }
         public decimal AverageObjectVolume { get; set; }
         public decimal VolumeStandardDeviation { get; set; }
         public decimal MinVolume { get; set; }
